# api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/go-gost/x/config
...
    NameserverConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            async:
                type: boolean
                x-go-name: Async
            chain:
                type: string
                x-go-name: Chain
            clientIP:
                type: string
                x-go-name: ClientIP
            hostname:
                type: string
                x-go-name: Hostname
            only:
                type: string
                x-go-name: Only
            prefer:
                type: string
                x-go-name: Prefer
            timeout:
                $ref: '#/definitions/Duration'
            ttl:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    PluginConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
            tls:
                $ref: '#/definitions/TLSConfig'
            token:
                type: string
                x-go-name: Token
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
...
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ResolverConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            nameservers:
                items:
                    $ref: '#/definitions/NameserverConfig'
                type: array
                x-go-name: Nameservers
            plugin:
                $ref: '#/definitions/PluginConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: github.com/go-gost/x/api
    RouterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            routes:
                items:
                    $ref: '#/definitions/RouterRouteConfig'
                type: array
                x-go-name: Routes
        type: object
        x-go-package: github.com/go-gost/x/config
    RouterRouteConfig:
        properties:
            gateway:
                type: string
                x-go-name: Gateway
            net:
                type: string
                x-go-name: Net
        type: object
        x-go-package: github.com/go-gost/x/config
    SDConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    TLSConfig:
        properties:
            caFile:
                type: string
                x-go-name: CAFile
            certFile:
                type: string
                x-go-name: CertFile
            commonName:
                type: string
                x-go-name: CommonName
            keyFile:
                type: string
                x-go-name: KeyFile
            options:
                $ref: '#/definitions/TLSOptions'
            organization:
                type: string
                x-go-name: Organization
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
            validity:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    TLSOptions:
        properties:
            cipherSuites:
                items:
                    type: string
                type: array
                x-go-name: CipherSuites
            maxVersion:
                type: string
                x-go-name: MaxVersion
            minVersion:
                type: string
                x-go-name: MinVersion
        type: object
        x-go-package: github.com/go-gost/x/config
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
...
    /config/resolvers:
        post:
            operationId: createResolverRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ResolverConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createResolverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new resolver, the name of the resolver must be unique in resolver list.
            tags:
                - Resolver
    /config/resolvers/{resolver}:
        delete:
            operationId: deleteResolverRequest
            parameters:
                - in: path
                  name: resolver
                  required: true
                  type: string
                  x-go-name: Resolver
            responses:
                "200":
                    $ref: '#/responses/deleteResolverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete resolver by name.
            tags:
                - Resolver
        put:
            operationId: updateResolverRequest
            parameters:
                - in: path
                  name: resolver
                  required: true
                  type: string
                  x-go-name: Resolver
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ResolverConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateResolverResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update resolver by name, the resolver must already exist.
            tags:
                - Resolver
...
    /config/routers:
        post:
            operationId: createRouterRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/RouterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createRouterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new router, the name of the router must be unique in router list.
            tags:
                - Router
    /config/routers/{router}:
        delete:
            operationId: deleteRouterRequest
            parameters:
                - in: path
                  name: router
                  required: true
                  type: string
                  x-go-name: Router
            responses:
                "200":
                    $ref: '#/responses/deleteRouterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete router by name.
            tags:
                - Router
        put:
            operationId: updateRouterRequest
            parameters:
                - in: path
                  name: router
                  required: true
                  type: string
                  x-go-name: Router
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/RouterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateRouterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update router by name, the router must already exist.
            tags:
                - Router
    /config/sds:
        post:
            operationId: createSDRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/SDConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createSDResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new SD, the name of the SD must be unique in SD list.
            tags:
                - SD
    /config/sds/{sd}:
        delete:
            operationId: deleteSDRequest
            parameters:
                - in: path
                  name: sd
                  required: true
                  type: string
                  x-go-name: SD
            responses:
                "200":
                    $ref: '#/responses/deleteSDResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete SD by name.
            tags:
                - SD
        put:
            operationId: updateSDRequest
            parameters:
                - in: path
                  name: sd
                  required: true
                  type: string
                  x-go-name: SD
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/SDConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateSDResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update SD by name, the SD must already exist.
            tags:
                - SD
...
produces:
    - application/json
responses:
...
    createResolverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createRouterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createSDResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteResolverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteRouterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteSDResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateResolverResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateRouterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateSDResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"